Moderate refactoring to use a background callback for output generation and to specify a different number of celery workers for output generation and downloading tasks.
